entrenar_modelo.py
"""
Script para entrenar modelo RandomForest (alternativo a XGBoost).
NOTA: Este script requiere funciones auxiliares que no están definidas.
Se recomienda usar train_model.py que es el script principal y está completo.
"""
# 🟦 M8. Entrenar modelo híbrido (RandomForest RANKING)

from sklearn.ensemble import RandomForestRegressor
import joblib
import sys

# NOTA: Este script está incompleto y requiere:
# - cargar_dataset() desde db.py o build_dataset.py
# - procesar() para feature engineering
# 
# Se recomienda usar train_model.py en su lugar que está completo y funcional.

def entrenar_modelo():
    """
    Entrena un modelo RandomForest para ranking.
    NOTA: Esta función requiere que cargar_dataset() y procesar() estén definidas.
    """
    print("⚠ Este script está incompleto y requiere funciones auxiliares.")
    print("   Se recomienda usar train_model.py en su lugar.")
    print("   Si deseas usar este script, necesitas implementar:")
    print("   - cargar_dataset(): función para cargar datos desde BD")
    print("   - procesar(): función para feature engineering")
    return False

if __name__ == "__main__":
    if not entrenar_modelo():
        sys.exit(1)
