requirements.txt
flask==3.1.2
flask-cors==6.0.1
pandas==2.3.3
numpy==2.3.5
scikit-learn==1.7.2
xgboost==3.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
joblib==1.5.2